+++ Shared/Entities/Models/GameSettings.cs
@@ -6,7 +6,7 @@ namespace Entities.Models
     {
         public int GameSettingId { get; set; }
         public int GameLength { get; set; }
-        public double WinningPot { get; set; }
+        public decimal WinningPot { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
     }
